Backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
pymongo==4.6.1
cloudinary==1.38.0
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
numpy==1.26.4
reportlab==4.1.0
python-dotenv==1.0.1
pydantic==2.6.1
torch>=2.0.0
torchvision>=0.15.0
requests>=2.31.0
